+++ SOC_Particle/src/Coulombs.h
@@ -39,22 +39,22 @@ class Coulombs
 {
 public:
   Coulombs();
-  Coulombs(double *sp_delta_q, const float q_cap_rated,
+  Coulombs(double *sp_delta_q, const double q_cap_rated,
     const double s_coul_eff, const float dx_voc, const float dy_voc, const float dz_voc);
   ~Coulombs();
-  void apply_cap_scale(const float scale);
+  void apply_cap_scale(const double scale);
   void apply_delta_q(const double delta_q);
   void apply_delta_q_t(const bool reset);
   void apply_delta_q_t(const double delta_q, const double tb_f);
   void apply_resetting(const bool resetting){ resetting_ = resetting; };
-  void apply_soc(const float soc, const double tb_f);
+  void apply_soc(const double soc, const double tb_f);
   void assign_all_mod() { chem_.assign_all_chm(); };
   double calculate_capacity(const double tb_f);
   Chemistry *chem() { return &chem_; };
   void chem_pretty_print () { chem_.pretty_print(); };
   double coul_eff() { return ( coul_eff_ ); };
   void coul_eff(const double coul_eff) { coul_eff_ = coul_eff; };
-  virtual float count_coulombs(Sensors *Sen, const bool reset_temp, const float charge_curr, const bool sat,
+  virtual double count_coulombs(Sensors *Sen, const bool reset_temp, const float charge_curr, const bool sat,
   const bool saturated);
   double d_delta_q() { return(d_delta_q_); };
   double delta_q() { return(*sp_delta_q_); };
@@ -68,15 +68,15 @@ public:
   void put_dy_voc(const float inp) { chem_.put_dy_voc(inp); }
   void put_dz_voc(const float inp) { chem_.put_dz_voc(inp); }
   double q(){ return (q_); };
-  float q_cap_rated(){ return (q_cap_rated_); };
-  float q_cap_rated_scaled(){ return (q_cap_rated_scaled_); };
-  float q_capacity(){ return (q_capacity_); };
+  double q_cap_rated(){ return (q_cap_rated_); };
+  double q_cap_rated_scaled(){ return (q_cap_rated_scaled_); };
+  double q_capacity(){ return (q_capacity_); };
   double q_inf(){ return (q_inf_); };
   bool sat() { return(sat_); };
   bool saturated() { return(saturated_); };
-  float soc() { return(soc_); };
-  float soc_inf() { return(soc_inf_); };
-  float soc_min() { return(soc_min_); };
+  double soc() { return(soc_); };
+  double soc_inf() { return(soc_inf_); };
+  double soc_min() { return(soc_min_); };
   double time_neg() { return(time_neg_); };
   double time_pos() { return(time_pos_); };
   virtual float vsat(void) = 0;
@@ -93,19 +93,19 @@ protected:
   double q_capacity_; // Saturation charge at temperature, C
   double q_cap_rated_;// Rated capacity at t_rated_, saved for future scaling, C
   double q_cap_rated_scaled_;// Applied rated capacity at t_rated_, after scaling, C
-  float q_inf_;       // Same as q_ but not reset on saturation, C
-  float q_min_;       // Floor on charge available to use, C
-  bool sat_;       // Indication that battery is potentially saturated, T=saturated
-  bool saturated_; // Battery is confirmed saturated, T=saturated
+  double q_inf_;      // Same as q_ but not reset on saturation, C
+  double q_min_;      // Floor on charge available to use, C
+  bool sat_;          // Indication that battery is potentially saturated, T=saturated
+  bool saturated_;    // Battery is confirmed saturated, T=saturated
   double soc_;        // Fraction of saturation charge (q_capacity_) available (0-1)
-  float soc_ekf_min_; // Minimum SOC for EKF operation
+  double soc_ekf_min_; // Minimum SOC for EKF operation
   double soc_inf_;    // Fraction of saturation charge (q_capacity_) available (-inf - inf)
-  float soc_min_;     // As battery cools, the voltage drops and there appears a minimum soc it can deliver
+  double soc_min_;    // As battery cools, the voltage drops and there appears a minimum soc it can deliver
   double *sp_delta_q_;// Charge since saturated, C
-  double tb_f_;        // Temperature, deg C
-  double tb_f_rate_;   // Tb rate, deg C / s
-  double Tb_f_;        // Temperature, deg C
-  double Tb_f_rate_;   // Tb rate, deg C / s
+  double tb_f_;       // Temperature, deg C
+  double tb_f_rate_;  // Tb rate, deg C / s
+  double Tb_f_;       // Temperature, deg C
+  double Tb_f_rate_;  // Tb rate, deg C / s
   double time_neg_;   // Time spent accumulating delta_q_neg_, s
   double time_pos_;   // Time spent accumulating delta_q_pos_, s
   Chemistry chem_;    // Chemistry
